+++ src/main/kotlin/days/Day3.kt
@@ -3,7 +3,7 @@ package days
 import Solution
 
 class Day3 : Solution {
-	override val day: Int = 3
+	override val day: UByte = 3u
 
 	override fun solve1(): String {
 		var sum: Int = 0
