+++ src/main/kotlin/days/Day9.kt
@@ -5,7 +5,7 @@ import kotlin.math.abs
 import kotlin.math.sign
 
 class Day9 : Solution {
-	override val day: Int = 9
+	override val day: UByte = 9u
 
 	private operator fun Pair<Int, Int>.plus(other: Pair<Int, Int>): Pair<Int, Int> {
 		return Pair(this.first + other.first, this.second + other.second)
